Remove unneeded import, fix bare except
The `os` module isn't needed, so we need not import it.

Additionally, a bare `except` clause is pretty much never desired, since
it includes all exceptions, including sigkills.

Instead, just check for an `ImportError`, since that's what we're really
trying to do: fall back on `distutils` if `setuptools` is not present.