Drop `mock` and `nose` as package dependencies
That `nose` is used as a test runner (and `mock` is used in those tests)
has nothing to do with the package itself. Rather, these are just
dependencies needed in order to *run tests.*

Note that we're still pinning to very precise version numbers, for no
particularly compelling reason. We'll fix that soon.

(`version` number is not increased, since this commit won't be packaged
& released).