Use plain assertions instead of preconditions
The `preconditions` library provides a convenience for simple checking
of arguments via a lambda function, but we can achieve essentially the
same thing by just using `assert` statements in the body of the
function.

Additionally, correct a typo in one of the preconditions, where we
assert than `a[0] <= a[0]` (which should be true for just about any type
in Python, and is certainly true for integers).